Time: 4 ms (9.32%), Space: 12.3 MB (10.59%) - LeetHub

+++ 0797-rabbits-in-forest/0797-rabbits-in-forest.cpp
@@ -17,7 +17,7 @@ public:
 
                 // if(num == 0) sum+=count;
                 
-                     double d = ceil((double)count / (num + 1));
+                     int d = ceil((double)count / (num + 1));
                      sum += d * (num+1);
            
             }
